+++ scripts/start_remote_mcp.ps1
@@ -69,6 +69,8 @@ $env:FREECAD_REQUIRE_BOUNDED_XMLRPC = 'true'
 $env:FREECAD_TRANSPORT = 'http'
 $env:FREECAD_HTTP_HOST = '127.0.0.1'
 $env:FREECAD_HTTP_PORT = '8000'
+$env:FREECAD_HTTP_JSON_RESPONSE = 'true'
+$env:FREECAD_HTTP_UNSTRUCTURED_TOOL_RESULTS = 'true'
 
 $existingListener = Get-NetTCPConnection `
     -LocalPort ([int]$env:FREECAD_HTTP_PORT) `